requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==21.2
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
